refactor(JsonEditor): connect JsonEditor file saving and dataset saving
