+++ src/Ml.Cli.WebApp/ClientApp/src/Annotation/AnnotationImagesLoader.js
@@ -28,13 +28,30 @@ const getImages = (fetchFunction) => async (item) => {
     return fetchImages(fetchResult);
 };
 
-const AnnotationImagesLoader = ({item, expectedOutput, onSubmit, MonacoEditor, parentState, fetchFunction}) => {
+const getHttpResultItem = async (item, fetchFunction) => {
+    const params = {
+        filePath: `${item.fileDirectory}\\${item.fileName}`
+    };
+    const fetchResult = await fetchGetData(fetchFunction)(params, "api/annotations");
+    if (fetchResult.status === 200) {
+        const dataObject = await fetchResult.json();
+        return {httpResult: dataObject, errorMessage: ""};
+    }
+    const errorMessage = `Requête invalide: ${fetchResult.statusText}`;
+    return {errorMessage, httpResult: {}};
+};
+
+const AnnotationImagesLoader = ({item, MonacoEditor, parentState, fetchFunction}) => {
 
     const [state, setState] = useState({
         fileUrls: [],
-        filePrimaryUrl: ""
+        filePrimaryUrl: "",
+        httpResultItem: {},
+        errorMessage: "",
+        isFetched: false
     });
-    
+
+    const mutationJson = useMutation(newData => fetchPostJson(fetchFunction)("/api/datasets/save", newData));
     const mutationDataset = useMutation(newData => fetchPostJson(fetchFunction)("/api/annotations/save", newData));
 
     const getUrls = async () => {
@@ -43,21 +60,47 @@ const AnnotationImagesLoader = ({item, expectedOutput, onSubmit, MonacoEditor, p
         return {fileUrls: newUrls, filePrimaryUrl: newFileUrl};
     };
 
+    const getEditorContent = async () => {
+        const {httpResult, errorMessage} = await getHttpResultItem(item, fetchFunction);
+        return {errorMessage, httpResultItem: httpResult, isFetched: true};
+    };
+
     useEffect(() => {
         let isMounted = true;
-        getUrls().then(obj => {
-            if(isMounted){
-                setState({fileUrls: obj.fileUrls, filePrimaryUrl: obj.filePrimaryUrl});
-            }
+        getUrls().then(urls => {
+            getEditorContent().then(content => {
+                if (isMounted) {
+                    setState({
+                        fileUrls: urls.fileUrls,
+                        filePrimaryUrl: urls.filePrimaryUrl,
+                        errorMessage: content.errorMessage,
+                        httpResultItem: content.httpResultItem,
+                        isFetched: content.isFetched
+                    });
+                }
+            })
         });
         return () => {
             isMounted = false;
         }
     }, []);
-    
+
+    const saveJsonEditor = editorContent => {
+        const newHttpResultItem = state.httpResultItem;
+        newHttpResultItem.body = editorContent;
+        mutationJson.mutate(newHttpResultItem);
+        setState({...state, httpResultItem: newHttpResultItem});
+    };
+
     const setAnnotationObject = e => {
         let returnedObject;
-        switch (parentState.annotationType){
+        switch (parentState.annotationType) {
+            case "JsonEditor":
+                saveJsonEditor(e);
+                returnedObject = {
+                    "content": e
+                };
+                break;
             case "Ocr":
                 returnedObject = {
                     "type": e.type,
@@ -77,7 +120,7 @@ const AnnotationImagesLoader = ({item, expectedOutput, onSubmit, MonacoEditor, p
         }
         return returnedObject;
     }
-    
+
     const onDatasetSubmit = e => {
         const annotationObject = {
             datasetLocation: parentState.datasetLocation,
@@ -87,31 +130,41 @@ const AnnotationImagesLoader = ({item, expectedOutput, onSubmit, MonacoEditor, p
         };
         mutationDataset.mutate(annotationObject);
     }
-    
+
     return (
         <>
             {parentState.annotationType === "JsonEditor" &&
-                <JsonEditorContainer
-                    expectedOutput={expectedOutput}
-                    urls={state.fileUrls}
-                    onSubmit={onSubmit}
-                    MonacoEditor={MonacoEditor}
-                />
+            <>
+                {!state.isFetched &&
+                    <div>Chargement de l'éditeur...</div>
+                }
+                {state.isFetched &&
+                    <JsonEditorContainer
+                        expectedOutput={{id: item.id, fileName: item.fileName, value: state.httpResultItem.body}}
+                        urls={state.fileUrls}
+                        onSubmit={onDatasetSubmit}
+                        MonacoEditor={MonacoEditor}
+                    />
+                }
+                {state.errorMessage &&
+                <div className="error-message">{state.errorMessage}</div>
+                }
+            </>
             }
             {parentState.annotationType === "Ocr" &&
-                <OcrLazy
-                    labels={parentState.configuration}
-                    expectedLabels={[]}
-                    url={state.filePrimaryUrl}
-                    onSubmit={onDatasetSubmit}
-                />
+            <OcrLazy
+                labels={parentState.configuration}
+                expectedLabels={[]}
+                url={state.filePrimaryUrl}
+                onSubmit={onDatasetSubmit}
+            />
             }
             {parentState.annotationType === "Cropping" &&
-                <CroppingLazy
-                    labels={parentState.configuration}
-                    url={state.filePrimaryUrl}
-                    onSubmit={onDatasetSubmit}
-                />
+            <CroppingLazy
+                labels={parentState.configuration}
+                url={state.filePrimaryUrl}
+                onSubmit={onDatasetSubmit}
+            />
             }
         </>
     );
