+++ src/Ml.Cli.WebApp/ClientApp/src/Annotation/TableAnnotateItem.js
@@ -1,54 +1,8 @@
-﻿import React, {useEffect, useState} from "react";
+﻿import React from "react";
 import AnnotationImagesLoader from "./AnnotationImagesLoader";
 import './TableAnnotateItem.scss';
-import {useMutation} from "react-query";
-import {fetchGetData, fetchPostJson} from "../FetchHelper";
-
-const getHttpResultItem = async (item, fetchFunction) => {
-    const params = {
-        filePath: `${item.fileDirectory}\\${item.fileName}`
-    };
-    const fetchResult = await fetchGetData(fetchFunction)(params, "api/annotations");
-    if (fetchResult.status === 200) {
-        const dataObject = await fetchResult.json();
-        return {httpResult: dataObject, errorMessage: ""};
-    }
-    const errorMessage = `Requête invalide: ${fetchResult.statusText}`;
-    return {errorMessage, httpResult: {}};
-};
 
 const TableAnnotateItem = ({parentState, item, MonacoEditor, fetchFunction}) => {
-    const [state, setState] = useState({
-        httpResultItem: {},
-        errorMessage: "",
-        isFetched: false
-    });
-
-    const mutationJson = useMutation(newData => fetchPostJson(fetchFunction)("/api/datasets/save", newData));
-
-    useEffect(() => {
-        let isMounted = true;
-        getEditorContent().then(obj => {
-            if(isMounted){
-                setState({...state, errorMessage: obj.errorMessage, httpResultItem: obj.httpResultItem, isFetched: obj.isFetched});
-            }
-        });
-        return () => {
-            isMounted = false;
-        }
-    }, []);
-
-    const getEditorContent = async () => {
-        const {httpResult, errorMessage} = await getHttpResultItem(item, fetchFunction);
-        return {errorMessage, httpResultItem: httpResult, isFetched: true};
-    };
-
-    const saveJson = editorContent => {
-        const newHttpResultItem = state.httpResultItem;
-        newHttpResultItem.body = editorContent;
-        mutationJson.mutate(newHttpResultItem);
-        setState({...state, httpResultItem: newHttpResultItem});
-    };
 
     return (
         <div className="table-result">
@@ -57,23 +11,12 @@ const TableAnnotateItem = ({parentState, item, MonacoEditor, fetchFunction}) => 
                     <p>Résultat fichier : {item.fileName}</p>
                 </div>
             </div>
-            {
-                !state.isFetched &&
-                <div>Chargement du fichier...</div>
-            }
-            {state.isFetched &&
             <AnnotationImagesLoader
                 item={item}
-                expectedOutput={{id: item.id, fileName: item.fileName, value: state.httpResultItem.body}}
-                onSubmit={saveJson}
                 MonacoEditor={MonacoEditor}
                 parentState={parentState}
                 fetchFunction={fetchFunction}
             />
-            }
-            {state.errorMessage &&
-            <div className="error-message">{state.errorMessage}</div>
-            }
         </div>
     );
 };
